<==== Mercurial data ====>
Branch: stable

+++ common/src/core/dataprovider/media_streamdataprovider.cpp
@@ -168,7 +168,7 @@ void QnAbstractMediaStreamDataProvider::checkTime(QnAbstractMediaDataPtr media)
         if (media->flags & QnAbstractMediaData::MediaFlags_BOF) {
             resetTimeCheck();
         }
-        else if (m_lastMediaTime[media->channelNumber] != AV_NOPTS_VALUE) {
+        else if ((quint64)m_lastMediaTime[media->channelNumber] != AV_NOPTS_VALUE) {
             qint64 timeDiff = media->timestamp - m_lastMediaTime[media->channelNumber];
         	// if timeDiff < -N it may be time correction or dayling time change
             if (timeDiff >=-TIME_RESYNC_THRESHOLD  && timeDiff < MIN_FRAME_DURATION)
